requirements.txt
flask
requests
azure-storage-blob
azure-identity
python-dotenv
azure.keyvault
jwcrypto
pytz
pandas